refactor: extract shared validators + support stdin for update
- Extract validateImportance, validateContentLength, MAX_CONTENT_LENGTH into src/validate.ts
- Reuse shared validators in store.ts and memory.ts (removes duplication)
- Add stdin support for update command (echo 'new' | memoclaw update <id>)
- Update tests to import from shared validate module

Closes MEM-176, MEM-177
