+++ src/commands/store.ts
@@ -2,22 +2,7 @@ import type { ParsedArgs } from '../args.js';
 import { request } from '../http.js';
 import { c } from '../colors.js';
 import { outputJson, out, success, info } from '../output.js';
-
-const MAX_CONTENT_LENGTH = 8192;
-
-function validateContentLength(content: string, label = 'Content') {
-  if (content.length > MAX_CONTENT_LENGTH) {
-    throw new Error(`${label} exceeds the ${MAX_CONTENT_LENGTH} character limit (got ${content.length} chars)`);
-  }
-}
-
-function validateImportance(value: string): number {
-  const n = parseFloat(value);
-  if (isNaN(n) || n < 0 || n > 1) {
-    throw new Error(`Importance must be a number between 0 and 1 (got "${value}")`);
-  }
-  return n;
-}
+import { validateContentLength, validateImportance } from '../validate.js';
 
 export async function cmdStore(content: string, opts: ParsedArgs) {
   validateContentLength(content);
